requirements1.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
astunparse==1.6.3
attrs==25.3.0
av==15.0.0
bitsandbytes==0.47.0
blessed==1.21.0
blinker==1.9.0
blosc2==3.6.1
boto3==1.40.10
botocore==1.40.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.3
cycler==0.12.1
datasets==4.0.0
deepspeed==0.17.4
dill==0.3.8
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
easydict==1.13
einops==0.8.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss==1.9.0
filelock==3.19.1
fitlog==0.9.15
flash_attn==2.8.3
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
gpustat==1.1.1
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.9
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.11.3
kiwisolver==1.4.9
langchain-core==0.3.75
langchain-openai==0.3.32
langsmith==0.4.21
libclang==18.1.1
lightning-utilities==0.15.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multimodal==0.0.13
multiprocess==0.70.16
namex==0.1.0
ndindex==1.10.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.1
numexpr==2.11.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.103.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
peft==0.17.1
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pytorch-ood==0.2.2
pytorch-pretrained-bert==0.6.2
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentence-transformers==5.1.0
sentry-sdk==2.34.1
setuptools==78.1.1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
tables==3.10.2
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
termcolor==3.1.0
tf_keras==2.20.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tiktoken==0.11.0
together==1.5.25
tokenizers==0.22.0
torch==2.8.0+cu126
torchmetrics==1.8.2
torchtext==0.18.0
torchvision==0.23.0+cu126
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
typeguard==4.4.4
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tyro==0.9.28
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.1
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
